This finishes the fix to log2. The computed dram size now matches the size indicated by byte 31 of SPD.
Memory is still not working; hanging in dqs training. 

Acked-by: Ronald G. Minnich <rminnich@gmail.com>



git-svn-id: svn://coreboot.org/repository/coreboot-v3@854 f3766cd6-281f-0410-b1cd-43a5c92072e9

+++ southbridge/amd/amd8111/acpi.c
@@ -185,8 +185,8 @@ static void acpi_read_resources(struct device * dev)
 	resource = new_resource(dev, 0x58);
 	resource->base  = 0;
 	resource->size  = 256;
-	resource->align = log2(256);
-	resource->gran  = log2(256);
+	resource->align = log2f(256);
+	resource->gran  = log2f(256);
 	resource->limit = 65536;
 	resource->flags = IORESOURCE_IO;
 	resource->index = 0x58;
